backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.9
Jinja2==3.1.4
fpdf2==2.7.9
google-generativeai>=0.3.0


